:twisted_rightwards_arrows: AST controller pattern (#34)
* Introduced controller logic for AST
* Upgraded GraPLHook4j version
* Improved readability and simplified AST methods
