+++ .gitignore
@@ -1,6 +1,5 @@
 /.gradle/
 /build/
-/coverage/
 
 # Ignore Gradle GUI config
 gradle-app.setting
