+++ src/main/java/za/ac/sun/grapl/visitors/ast/ASTClassVisitor.java
@@ -15,35 +15,24 @@
  */
 package za.ac.sun.grapl.visitors.ast;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
-import za.ac.sun.grapl.domain.models.vertices.FileVertex;
-import za.ac.sun.grapl.domain.models.vertices.NamespaceBlockVertex;
-import za.ac.sun.grapl.hooks.IHook;
-
-import java.util.Objects;
+import za.ac.sun.grapl.controllers.ASTController;
 
 public class ASTClassVisitor extends ClassVisitor implements Opcodes {
 
-    final static Logger logger = LogManager.getLogger();
+    private final ASTController controller;
 
-    final IHook hook;
-    private String classPath;
-    private int order;
-    private FileVertex fv;
-    private ASTMethodVisitor astMv;
-
-    public ASTClassVisitor(IHook hook, ClassVisitor cv) {
+    public ASTClassVisitor(ClassVisitor cv) {
         super(ASM5, cv);
-        this.hook = hook;
-        this.order = 0;
+        this.controller = ASTController.getInstance();
+        this.controller.resetOrder();
     }
 
     @Override
     public void visit(int version, int access, String name, String signature, String superName, String[] interfaces) {
+        super.visit(version, access, name, signature, superName, interfaces);
         String className;
         String namespace;
         if (name.lastIndexOf('/') != -1) {
@@ -53,66 +42,22 @@ public class ASTClassVisitor extends ClassVisitor implements Opcodes {
             className = name;
             namespace = "";
         }
-        this.classPath = name.replaceAll("/", ".");
         namespace = namespace.replaceAll("/", ".");
 
-        // Build NAMESPACE_BLOCK if packages are present
-        NamespaceBlockVertex nbv = null;
-        if (!namespace.isEmpty()) {
-            // Populate namespace block chain
-            String[] namespaceList = namespace.split("\\.");
-            if (namespaceList.length > 0) nbv = this.populateNamespaceChain(namespaceList);
-        }
-
-        fv = new FileVertex(className, order++);
-        // Join FILE and NAMESPACE_BLOCK if namespace is present
-        if (!Objects.isNull(nbv)) {
-            this.hook.joinFileVertexTo(fv, nbv);
-        }
-
+        controller.projectFileAndNamespace(namespace, className);
         // TODO: Could create MEMBER vertex from here to declare member classes
+    }
 
-        super.visit(version, access, name, signature, superName, interfaces);
+    @Override
+    public MethodVisitor visitMethod(int access, String name, String descriptor, String signature, String[] exceptions) {
+        controller.pushMethod(name, descriptor, access);
+        MethodVisitor mv = super.visitMethod(access, name, descriptor, signature, exceptions);
+        return new ASTMethodVisitor(mv);
     }
 
     @Override
     public void visitEnd() {
         super.visitEnd();
-    }
-
-    @Override
-    public MethodVisitor visitMethod(int access, String name, String descriptor, String signature, String[] exceptions) {
-        order = astMv == null ? order : astMv.getOrder();
-        MethodVisitor mv = super.visitMethod(access, name, descriptor, signature, exceptions);
-        astMv = new ASTMethodVisitor(mv, hook, access, name, classPath, descriptor, fv);
-        astMv.setOrder(order);
-        return astMv;
-    }
-
-    /**
-     * Creates a change of namespace block vertices and returns the final one in the chain.
-     *
-     * @param namespaceList a list of package names
-     * @return the final namespace block vertex in the chain (the one associated with the file)
-     */
-    private NamespaceBlockVertex populateNamespaceChain(String[] namespaceList) {
-        NamespaceBlockVertex prevNamespaceBlock = new NamespaceBlockVertex(namespaceList[0], namespaceList[0], order++);
-        if (namespaceList.length == 1) return prevNamespaceBlock;
-
-        NamespaceBlockVertex currNamespaceBlock = null;
-        StringBuilder namespaceBuilder = new StringBuilder(namespaceList[0]);
-        for (int i = 1; i < namespaceList.length; i++) {
-            namespaceBuilder.append(".".concat(namespaceList[i]));
-            currNamespaceBlock = new NamespaceBlockVertex(namespaceList[i], namespaceBuilder.toString(), order++);
-            this.hook.joinNamespaceBlocks(prevNamespaceBlock, currNamespaceBlock);
-            prevNamespaceBlock = currNamespaceBlock;
-        }
-        return currNamespaceBlock;
-    }
-
-    public ASTClassVisitor order(int order) {
-        this.order = order;
-        return this;
     }
 
 }
